requirements.txt
requests
PyYAML
python-dotenv
schedule
pandas